+++ test/uk/gov/hmrc/eeitt/repositories/EnrolmentRepositorySpec.scala
@@ -2,24 +2,19 @@ package uk.gov.hmrc.eeitt.repositories
 
 import org.scalatest.{ BeforeAndAfterEach, Inspectors, LoneElement }
 import org.scalatest.concurrent.{ IntegrationPatience, ScalaFutures }
-import reactivemongo.bson.BSONObjectID
+import uk.gov.hmrc.eeitt.RepositorySupport
 import uk.gov.hmrc.eeitt.model.Enrolment
-import uk.gov.hmrc.mongo.MongoSpecSupport
 import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.ExecutionContext.Implicits.global
 
-class EnrolmentRepositorySpec extends UnitSpec with MongoSpecSupport with BeforeAndAfterEach with ScalaFutures with LoneElement with Inspectors with IntegrationPatience {
-
-  val repo = new EnrolmentRepository
-  val fakeId = BSONObjectID.generate
+class EnrolmentRepositorySpec extends UnitSpec with RepositorySupport with BeforeAndAfterEach with ScalaFutures with LoneElement with Inspectors with IntegrationPatience {
 
   override protected def beforeEach(): Unit = {
     await(repo.removeAll())
   }
 
   "query all enrolments" should {
-
     "produce all enrolments from repository" in {
       insertEnrolment(Enrolment(fakeId, "1", "12LT31", true, "SE39EP"))
       insertEnrolment(Enrolment(fakeId, "2", "12LT32", true, "SE39XY"))
@@ -27,35 +22,24 @@ class EnrolmentRepositorySpec extends UnitSpec with MongoSpecSupport with Before
       await(repo.count) shouldBe 3
       await(repo.getAllEnrolments()).size shouldBe 3
     }
-
   }
 
   "query enrolments with a form type" should {
-
     "produce all enrolments from repository with a given form type" in {
       insertEnrolment(Enrolment(fakeId, "1", "12LT34", true, "SE39EP"))
       insertEnrolment(Enrolment(fakeId, "1", "12LT35", true, "SE39XY"))
       await(repo.count) shouldBe 2
       await(repo.getAllEnrolmentsWithFormType("1")).size shouldBe 2
     }
-
   }
 
   "lookup enrolments by registration number " should {
-
     "find enrolment with a given registration number" in {
       insertEnrolment(Enrolment(fakeId, "1", "12LT34", true, "SE39EP"))
       insertEnrolment(Enrolment(fakeId, "1", "12LT35", true, "SE39XY"))
       await(repo.count) shouldBe 2
       await(repo.lookupEnrolment("12LT35")).size shouldBe 1
     }
-
-  }
-
-  def insertEnrolment(enrolment: Enrolment): BSONObjectID = {
-    val e = Enrolment(_id = BSONObjectID.generate, enrolment.formTypeRef, enrolment.registrationNumber, enrolment.livesInTheUk, enrolment.postcode)
-    await(repo.collection.insert(e))
-    e._id
   }
 
 }
