+++ test/uk/gov/hmrc/eeitt/validation/LookupTestCases.scala
@@ -2,23 +2,19 @@ package uk.gov.hmrc.eeitt.validation
 
 import org.scalatest.{ BeforeAndAfterEach, Inspectors, LoneElement }
 import org.scalatest.concurrent.{ IntegrationPatience, ScalaFutures }
-import reactivemongo.bson.BSONObjectID
-import uk.gov.hmrc.eeitt.model.EnrolmentVerificationResponse.{RESPONSE_OK,RESPONSE_NOT_FOUND,INCORRECT_POSTCODE,INCORRECT_REGIME}
+import uk.gov.hmrc.eeitt.RepositorySupport
+import uk.gov.hmrc.eeitt.model.EnrolmentVerificationResponse.{ INCORRECT_POSTCODE, INCORRECT_REGIME, RESPONSE_NOT_FOUND, RESPONSE_OK }
 import uk.gov.hmrc.eeitt.model.{ Enrolment, EnrolmentVerificationRequest, EnrolmentVerificationResponse }
-import uk.gov.hmrc.eeitt.repositories.EnrolmentRepository
 import uk.gov.hmrc.eeitt.services.EnrolmentVerificationService
-import uk.gov.hmrc.mongo.MongoSpecSupport
 import uk.gov.hmrc.play.test.UnitSpec
 
 import scala.concurrent.ExecutionContext.Implicits.global
+
 /**
  * Created by harrison on 17/10/16.
- * Exyended by milosz on 18/10/16.
+ * Extended by milosz on 18/10/16.
  */
-class LookupTestCases extends UnitSpec with MongoSpecSupport with BeforeAndAfterEach with ScalaFutures with LoneElement with Inspectors with IntegrationPatience {
-
-  val repo = new EnrolmentRepository
-  val fakeId = BSONObjectID.generate
+class LookupTestCases extends UnitSpec with RepositorySupport with BeforeAndAfterEach with ScalaFutures with LoneElement with Inspectors with IntegrationPatience {
 
   object TestEnrolmentVerificationService extends EnrolmentVerificationService {
     val enrolmentRepo = repo
@@ -33,8 +29,8 @@ class LookupTestCases extends UnitSpec with MongoSpecSupport with BeforeAndAfter
   "Look up enrollments by registration number" should {
 
     /**
-      * Cases 7 and 11
-      */
+     * Cases 7 and 11
+     */
     "produce 'registration number incorrect' response when enrolment is not found" in {
       insertEnrolment(Enrolment(fakeId, "Aggregate Levy", "AL1111111111111", true, "ME1 9AB"))
       repo.count.futureValue shouldBe 1
@@ -43,9 +39,9 @@ class LookupTestCases extends UnitSpec with MongoSpecSupport with BeforeAndAfter
     }
 
     /**
-      * Cases 8 and 12
-      */
-    "produce wrong postcode response when stored postcode is different than requested postcode" in {
+     * Cases 8 and 12
+     */
+    "produce 'incorrect postcode' response when stored postcode is different than requested postcode" in {
       insertEnrolment(Enrolment(fakeId, "Aggregate Levy", "AL9876543210123", true, "ME1 9AB"))
       repo.count.futureValue shouldBe 1
       val response = service.verify(EnrolmentVerificationRequest("Aggregate Levy", "AL9876543210123", true, "N12 6FG"))
@@ -59,8 +55,8 @@ class LookupTestCases extends UnitSpec with MongoSpecSupport with BeforeAndAfter
     }
 
     /**
-      * Cases 9 and 13
-      */
+     * Cases 9 and 13
+     */
     "produce 'regime incorrect' response if stored regime is different than requested regime" in {
       insertEnrolment(Enrolment(fakeId, "Aggregate Levy", "AL9876543210123", true, "ME1 9AB"))
       repo.count.futureValue shouldBe 1
@@ -69,19 +65,13 @@ class LookupTestCases extends UnitSpec with MongoSpecSupport with BeforeAndAfter
     }
 
     /**
-      * Cases 10 and 14
-      */
+     * Cases 10 and 14
+     */
     "produce successful response when enrolment is found and passed the validation" in {
       insertEnrolment(Enrolment(fakeId, "Aggregate Levy", "AL9876543210123", true, "ME1 9AB"))
       repo.count.futureValue shouldBe 1
       val response = service.verify(EnrolmentVerificationRequest("Aggregate Levy", "AL9876543210123", true, "ME1 9AB"))
       response.futureValue shouldBe EnrolmentVerificationResponse(RESPONSE_OK)
     }
-  }
-
-  def insertEnrolment(enrolment: Enrolment): BSONObjectID = {
-    val lease = Enrolment(_id = BSONObjectID.generate, enrolment.formTypeRef, enrolment.registrationNumber, enrolment.livesInTheUk, enrolment.postcode)
-    await(repo.collection.insert(lease))
-    lease._id
   }
 }
